chore(tasks): refactor prover deploy tasks
